third_party/blink/web_tests/external/wpt/editing/run/formatblock_3001-4000-expected.txt

This is a testharness.js-based test.
Found 87 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<h5>"]] "<p>[foobar]</p>" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "h5" but got "p"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<address>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<address>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<article>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<aside>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<blockquote>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<body>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<body>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<dd>"]] "<p>[foo<p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd>foo</dd><dd>bar</dd></dl>" but got "<dd>foo<br>bar</dd>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<dd>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<dd>"]] "<p>[foo<p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dd>foo</dd><dd>bar</dd></dl>" but got "<dd>foo<br>bar</dd>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<dd>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<del>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<del>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<details>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<details>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<dir>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<dir>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<dl>"]] "<p>[foo<p>bar]": execCommand("formatblock", false, "<dl>") return value
  assert_equals: expected false but got true
[FAIL] [["formatblock","<dl>"]] "<p>[foo<p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "<dl>foo<br>bar</dl>"
[FAIL] [["formatblock","<dl>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<dl>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got "dl"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<dt>"]] "<p>[foo<p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dt>bar</dt></dl>" but got "<dt>foo<br>bar</dt>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<dt>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<dt>"]] "<p>[foo<p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<dl><dt>foo</dt><dt>bar</dt></dl>" but got "<dt>foo<br>bar</dt>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<dt>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<fieldset>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<fieldset>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<figcaption>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<figcaption>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<figure>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<figure>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<footer>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<form>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<form>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<h1>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<h1>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<h2>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<h2>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<h3>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<h3>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<h4>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<h4>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<h5>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<h5>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<h6>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<h6>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<header>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<head>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<head>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<hgroup>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<hr>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<hr>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<html>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<html>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<ins>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<ins>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<li>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<li>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<listing>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<listing>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<menu>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<menu>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<nav>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<ol>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<ol>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<p>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<p>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<plaintext>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<plaintext>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<pre>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<pre>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<section>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<ul>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<ul>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<xmp>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<xmp>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<quasit>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") before
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["formatblock","<quasit>"]] "<p>[foo<p>bar]" queryCommandValue("formatblock") after
  assert_equals: Wrong result returned expected "p" but got ""
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<div>"]] "<ol><li>[foo]<li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><div>foo</div></li><li>bar</li></ol>" but got "<div><ol><li>foo</li></ol></div><ol><li>bar</li></ol>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<div>"]] "<ol><li>[foo]<li>bar</ol>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li><div>foo</div></li><li>bar</li></ol>" but got "<div><ol><li>foo</li></ol></div><ol><li>bar</li></ol>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<h1>[foo]<br>bar</h1>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><h1>bar</h1>" but got "<h1><p>foo</p>bar</h1>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<h1>[foo]<br>bar</h1>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><h1>bar</h1>" but got "<h1><p>foo</p>bar</h1>"
[FAIL] [["defaultparagraphseparator","div"],["formatblock","<p>"]] "<h1>foo<br>[bar]</h1>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<h1>foo</h1><p>bar</p>" but got "<h1>foo<br><p>bar</p></h1>"
[FAIL] [["defaultparagraphseparator","p"],["formatblock","<p>"]] "<h1>foo<br>[bar]</h1>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<h1>foo</h1><p>bar</p>" but got "<h1>foo<br><p>bar</p></h1>"
Harness: the test ran to completion.
